requirements.txt
absl-py==0.1.13
astor==0.6.2
backports.weakref==1.0.post1
bleach==1.5.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
grpcio==1.10.1
html5lib==0.9999999
Markdown==2.6.11
mock==2.0.0
numpy==1.14.2
OpenNMT-tf==1.7.0
pbr==4.0.1
pkg-resources==0.0.0
protobuf==3.6.1
pyonmttok==1.6.1
PyYAML==3.12
rouge==0.3.1
six==1.11.0
tensorboard==1.7.0
tensorflow-gpu==1.4.1
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
Werkzeug==0.14.1
